requirements.txt
pillow>=10.0.0
moviepy>=1.0.3
PyYAML>=6.0.1
